+++ Backend/app/src/main/java/com/example/app/miscellaneous/Workload.java
@@ -2,39 +2,39 @@ package com.example.app.miscellaneous;
 
 public class Workload {
 
-    private Long rating;
+    private Float rating;
 
-    private Long difficulty;
+    private Float difficulty;
 
-    private Long time_commitment;
+    private Float time_commitment;
 
-    public Workload(Long rating, Long difficulty, Long time_commitment) {
+    public Workload(Float rating, Float difficulty, Float time_commitment) {
         this.rating = rating;
         this.difficulty = difficulty;
         this.time_commitment = time_commitment;
     }
 
-    public Long getRating() {
+    public Float getRating() {
         return rating;
     }
 
-    public void setRating(Long rating) {
+    public void setRating(Float rating) {
         this.rating = rating;
     }
 
-    public Long getDifficulty() {
+    public Float getDifficulty() {
         return difficulty;
     }
 
-    public void setDifficulty(Long difficulty) {
+    public void setDifficulty(Float difficulty) {
         this.difficulty = difficulty;
     }
 
-    public Long getTime_commitment() {
+    public Float getTime_commitment() {
         return time_commitment;
     }
 
-    public void setTime_commitment(Long time_commitment) {
+    public void setTime_commitment(Float time_commitment) {
         this.time_commitment = time_commitment;
     }
 }
